@@ -4,6 +4,8 @@
 import React from 'react';
 import { storiesOf, action, addDecorator } from '@kadira/storybook';
 import { muiTheme } from 'storybook-addon-material-ui';
+import { WithNotes } from '@kadira/storybook-addon-notes';
+import { withKnobs, text, boolean, number } from '@kadira/storybook-addon-knobs';
 
 import App from '../App';
 import Header from '../Header';
@@ -22,22 +24,54 @@ reqThemes.keys().forEach((filename) => {
  *  You can pass a single object or an array of themes
  */
 
+
 storiesOf('React App', module)
+    .addDecorator(muiTheme(themesList)) /* [lightTheme, darkTheme, greyTheme]*/
     .addDecorator(story => (
       <div style={{ display: 'flex', justifyContent: 'center' }}>
-        <div style={{ width: '100%', /* maxWidth: 500,*/ minWidth: 200 }}>
+        <div style={{ width: '50%', maxWidth: 500, minWidth: 200 }}>
           {story()}
           {/* SHOW_SUPPORT ? <SupportProject /> : null*/}
         </div>
       </div>
     ))
-    .addDecorator(muiTheme(themesList)) /* [lightTheme, darkTheme, greyTheme]*/
+    .addDecorator(withKnobs)
     .add('App', () => (
       <App />
     ))
-    .add('App-header', () => (
-      <Header />
+    .addWithInfo('App-header', '<Header />', () => withNote(
+      `
+        Header Component
+
+        source: src/Header.jsx
+        story: src/stories
+        test: src/tests
+      `,
+      <Header
+        title={text('Title', 'Welcome to React-Theming')}
+        subtitle={text('Subtitle', 'Storybook Boilerplate Project')}
+      />,
     ))
-    .add('App-intro', () => (
-      <Intro />
+    .addWithInfo('App-intro', '<Intro />', () => withNote(
+      `
+        Intro Component
+
+        source: src/Intro.jsx
+        story: src/stories
+        test: src/tests
+      `,
+      <Intro />,
+    ))
+    .add('Addons Knobs & Notes', () => (
+      <div>
+        {text('Label1', 'Hello Button')}
+        {withNote(text('Label2', 'Hello Button'))}
+      </div>
     ));
+
+
+function withNote(note, child) {
+    return (
+      <WithNotes notes={note}>{child || null}</WithNotes>
+    );
+}
